cleaned training/validation
