+++ utils.py
@@ -1,4 +1,6 @@
 import numpy as np
+import torch
+import os
 
     # creates wordPairs for the Training & Validation Set
 def createWordPairs(indexes, tokens, window_size):
@@ -22,9 +24,6 @@ def createWordPairs(indexes, tokens, window_size):
                 idx_pairs.append((indices[center_word_pos], context_word_idx))
     return np.array(idx_pairs)
 
-    
-
-
 class AverageMeter():
     """Computes and stores the average and current value"""
     def __init__(self):
@@ -43,3 +42,47 @@ class AverageMeter():
         self.sum += val * n
         self.count += n
         self.avg = self.sum / self.count
+        
+        
+class EarlyStopping:
+    """Early stops the training if validation loss doesn't improve after a given patience."""
+    def __init__(self, patience, verbose, save_name):
+        """
+        Args:
+            patience (int): How long to wait after last time validation loss improved.
+                            Default: 7
+            verbose (bool): If True, prints a message for each validation loss improvement. 
+                            Default: False
+        """
+        self.patience = patience
+        self.verbose = verbose
+        self.counter = 0
+        self.best_score = None
+        self.early_stop = False
+        self.val_loss_min = np.Inf
+        self.save_name = save_name
+
+    def __call__(self, val_loss, model):
+
+        score = -val_loss
+
+        if self.best_score is None:
+            self.best_score = score
+            self.save_checkpoint(val_loss, model)
+        elif score < self.best_score:
+            self.counter += 1
+            print(f'EarlyStopping counter: {self.counter} out of {self.patience}')
+            if self.counter >= self.patience:
+                self.early_stop = True
+        else:
+            self.best_score = score
+            self.save_checkpoint(val_loss, model)
+            self.counter = 0
+
+    def save_checkpoint(self, val_loss, model):
+        '''Saves model when validation loss decrease.'''
+        if self.verbose:
+            print(f'Validation loss decreased ({self.val_loss_min:.6f} --> {val_loss:.6f}).  Saving model ...')
+        torch.save(model.input_embeddings(), os.path.join("SOFTMAX_trained", self.save_name))
+        self.val_loss_min = val_loss
+
